requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.3
requests==2.31.0